% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lstm_keras.R
\name{fit_lstm_keras}
\alias{fit_lstm_keras}
\title{Function to fit LSTM recurrent neural net in RTMB}
\usage{
fit_lstm_keras(data, hidden_dim = 5, n_proj_years = 2, last_year = NA)
}
\arguments{
\item{data}{data.frame with weight, year, and age}

\item{hidden_dim}{dimension of hidden layers (nhidden_layer + 2 for in/out)}

\item{nhidden_layer}{number of hidden layers (this is on a per obs basis)}
}
\description{
Function to fit LSTM recurrent neural net in RTMB
}
