% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lstm_tmb.r
\name{fit_lstm_rtmb}
\alias{fit_lstm_rtmb}
\title{Function to fit LSTM recurrent neural net in RTMB}
\usage{
fit_lstm_rtmb(data, nhidden_layer = 1, hidden_dim = 1, input_par = NULL)
}
\arguments{
\item{data}{data.frame with weight, year, and age}

\item{nhidden_layer}{number of hidden layers (this is on a per obs basis)}

\item{hidden_dim}{dimension of hidden layers (nhidden_layer + 2 for in/out)}
}
\description{
Function to fit LSTM recurrent neural net in RTMB
}
