% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nn_tmb.R
\name{fit_nn_rtmb}
\alias{fit_nn_rtmb}
\title{Function to fit neural net in RTMB}
\usage{
fit_nn_rtmb(data, nhidden_layer = 3, hidden_dim = 5, input_par = NULL)
}
\arguments{
\item{data}{data.frame with weight, year, and age}

\item{nhidden_layer}{number of hidden layers}

\item{hidden_dim}{dimension of hidden layers}
}
\description{
Function to fit neural net in RTMB
}
