% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gmrf.R
\name{FitGMRF}
\alias{FitGMRF}
\title{Function to fit GMRF weight-at-age model from Cheng et al 2023}
\usage{
FitGMRF(data = NULL, weights = NULL, n_proj_years = 2, last_year = NA)
}
\arguments{
\item{data}{data.frame with columns "year", "age", and "weight". Weight should be in kg so that uncertainty calculations do not cause an issue.}

\item{weights}{likelihood weights for model.}

\item{n_proj_years}{the number of years to project forward}
}
\value{
a list of GMRF models
}
\description{
https://doi.org/10.1016/j.fishres.2023.106755
}
