% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gmrf_rtmb.R
\name{FitGMRF_RTMB}
\alias{FitGMRF_RTMB}
\title{Function to fit GMRF weight-at-age model from Cheng et al 2023
https://github.com/chengmatt/GMRF_WAA/blob/master/R_scripts/model%20runs/RTMB_GrowthModel.R}
\usage{
FitGMRF_RTMB(
  data = NULL,
  weights = NULL,
  n_proj_years = 2,
  last_year = NA,
  n.newton = 3
)
}
\arguments{
\item{data}{data.frame with columns "year", "age", and "weight". Weight should be in kg so that uncertainty calculations do not cause an issue.}

\item{weights}{likelihood weights for model.}

\item{n_proj_years}{the number of years to project forward}

\item{n.newton}{number of extra newton steps we want to take}
}
\value{
a list of GMRF models
}
\description{
https://doi.org/10.1016/j.fishres.2023.106755
}
