% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vbgf.R
\name{FitVBGF}
\alias{FitVBGF}
\title{Fit 3-parameter von-bertalanfy growth function}
\usage{
FitVBGF(data, weights = NULL, n_proj_years = 2)
}
\arguments{
\item{n_proj_years}{number of years to project forward (default = 2)}
}
\description{
Fit 3-parameter von-bertalanfy growth function
}
