% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wtagere.R
\name{FitWtAgeRE}
\alias{FitWtAgeRE}
\title{Function to fit weight-at-age model from Jim Ianelli}
\usage{
FitWtAgeRE(data = NULL, weights = NULL, n_proj_years = 2, last_year = NULL)
}
\arguments{
\item{data}{data.frame with columns "year", "age", and "weight". Weight should be in kg so that uncertainty calculations do not cause an issue.}

\item{weights}{likelihood weights for model.}

\item{n_proj_years}{the number of years to project forward}
}
\value{
a list models
}
\description{
https://github.com/afsc-assessments/WtAgeRe/tree/main
}
