% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xgboost.R
\name{FitXGBoost}
\alias{FitXGBoost}
\title{Function to fit XGBoost (eXtreme Gradient Boosting) to weight-at-age data}
\usage{
FitXGBoost(
  form = NULL,
  data = NULL,
  weights = NULL,
  n_proj_years = 2,
  max.depth = 6,
  nthread = 1,
  nrounds = 2
)
}
\arguments{
\item{form}{model formula}

\item{data}{data.frame with columns "year", "age", and "weight" and other potential covariates. Weight should be in kg so that uncertainty calculations do not cause an issue.}

\item{weights}{likelihood weights for model.}

\item{n_proj_years}{the number of years to project forward}

\item{max.depth}{maximum depth of a tree. Default: 6}

\item{nthread}{description}

\item{nrounds}{max number of boosting iterations. Default: 2}
}
\value{
an XGBoost model object
}
\description{
This fits a basic XGBoost model to weight-at-age data using squared log error. The model specification can be optimized for the data-set. See ?xgboost for more information
https://xgboost.readthedocs.io/en/stable/R-package/xgboostPresentation.html
}
