% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastGrowth.R
\name{ForecastGrowth}
\alias{ForecastGrowth}
\title{Forecast growth}
\usage{
ForecastGrowth(
  form = formula(weight ~ age + year),
  data = NULL,
  n_proj_years = 2,
  peels = 3
)
}
\arguments{
\item{form}{model formula for machine learning based methods and selection of covariates}

\item{data}{data.frame with columns "year", "age", and "weight" and any covariates. Weight should be in kg so that uncertainty calculations do not cause an issue.}

\item{n_proj_years}{the max number of years to project forward (default = 2)}

\item{peels}{the number of years to peel back for retrospective forecast skill evaluation (default = 3)}
}
\value{
A list with:
* all_predictions = data.frame of observed and model predicted (hindcasts & forecasts) weight-at-age for each model and retrospective peel
* all_forecasts = data.frame of forecasted weight-at-age for each model and retrospective peel
* rmse_table = data.frame of root mean squared error across all ages/peels from retrospective forecasts
* rmse_table_by_age = data.frame of root mean squared error by age across peels from retrospective forecasts
* best_mods = data.frame with of the best model from\code{rmse_table}
* best_mods_by_age = data.frame with of the best model from\code{rmse_table_by_age} for each age
* best_forecast = data.frame with forecasted weight-at-age from\code{best_mods}
* plots = list of plots from each model
}
\description{
The function takes a \code{data.frame} of weight-at-age data and compares the retrospective forecast skill of a variety of weight-at-age models. This is done by peeling the data \code{peels} times and comparing the observed vs forecasted data using root mean squared error (RMSE) across peels and all ages (or by each age) for each of the projection years set by \code{n_proj_years}. The function then outputs the forecasted weight-at-age based on the model with the lowest RMSE for each projection year.
}
\details{
Fits various weight-at-age models given input data and evaluates forecast skill using retrospective forecasting. Outputs the forecasted weight-at-age from the "best" performing model.
}
