% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastGrowth.R
\name{ForestGrowth}
\alias{ForestGrowth}
\title{Forecast growth}
\usage{
ForestGrowth(
  form = formula(weight ~ age + year),
  data = NULL,
  n_proj_years = 2,
  peels = 3
)
}
\arguments{
\item{form}{model formula for machine learning based methods and selection of covariates}

\item{data}{data.frame with columns "year", "age", and "weight" and any covariates. Weight should be in kg so that uncertainty calculations do not cause an issue.}

\item{n_proj_years}{the max number of years to project forward (default = 2)}

\item{weights}{likelihood weights for model.}
}
\description{
Fits various weight-at-age models given input data
}
