% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_growthPars.R
\name{get_growthPars}
\alias{get_growthPars}
\title{Obtain and format growth parameters (mean and covariance) from FishLife}
\usage{
get_growthPars(species_name)
}
\arguments{
\item{species_name}{one of "pop", "pcod", "flathead" OR latin genus-species}
}
\description{
Obtain and format growth parameters (mean and covariance) from FishLife
}
