% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthForecast-package.R
\docType{package}
\name{GrowthForecast-package}
\alias{GrowthForecast}
\alias{GrowthForecast-package}
\title{GrowthForecast: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
