% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lstm_rtmb.r
\name{lstm_neuron}
\alias{lstm_neuron}
\title{LSTM Cell Function}
\usage{
lstm_neuron(x, h_prev, c_prev, input_weights, hidden_weights, bias_vector)
}
\arguments{
\item{bias_vector}{}
}
\description{
LSTM Cell Function
}
