% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_FishLife.R
\name{reformat_FishLife}
\alias{reformat_FishLife}
\title{Bias correct mean and covariance values from FishLife (utility function)}
\usage{
reformat_FishLife(predict = predict0, trait = "Loo")
}
\arguments{
\item{predict}{list object from of FishLife::PlotTaxa}

\item{trait}{specific trait eg K, M, Winfinity, Loo}
}
\description{
Bias correct mean and covariance values from FishLife (utility function)
}
