% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wtagere.R
\name{WtAgeRE}
\alias{WtAgeRE}
\title{WtAgeRE in RTMB
Adopted from Jim Ianelli for RTMB
https://github.com/afsc-assessments/WtAgeRe}
\usage{
WtAgeRE(pars, data_list)
}
\arguments{
\item{data_list}{}
}
\description{
WtAgeRE in RTMB
Adopted from Jim Ianelli for RTMB
https://github.com/afsc-assessments/WtAgeRe
}
