% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_weight}
\alias{simulate_weight}
\title{Simulate weight-at-age data}
\usage{
simulate_weight(
  nyrs = 10,
  nsamples = 1000,
  nages = 20,
  winf = 5,
  k = 0.3,
  t0 = -0.5,
  sigma_obs = 0.05,
  sigma.year = c(0.05, 0.05, 0.2),
  rho = 0.3,
  rho_ar1 = 0.95,
  trend_beta = 0,
  seed = 1234
)
}
\arguments{
\item{nyrs}{Number of years}

\item{nsamples}{Number of samples across years and ages}

\item{nages}{Number of ages (starting at 1)}

\item{winf}{Mean asymptoptic weight}

\item{k}{Mean growth coefficient}

\item{t0}{Mean t0}

\item{sigma_obs}{Variance of lognormal observation error}

\item{sigma.year}{Vector of length 3 of MVN variances for von Bertalanfy parameter annual deviates (Winf, K, t0)}

\item{rho}{Correlation for von Bertalanfy growth parameters (Winf, K, t0) MVN annual deviates}

\item{rho_ar1}{Vector of length 3 of AR1 correlation for von Bertalanfy growth parameters (Winf, K, t0)}

\item{seed}{}
}
\description{
This code simulates weight-at-age data for individuals from a population where there are multiple years
based on a von Bertalanfy growth curve with environmental processes on the parameters:

\deqn{Weight_i = Winf_{yr[i]} * (1 - exp(k_{yr[i]} * (age_i - t0_{yr[i]})))}

\deqn{VGBM_{yr} = AR1 * VGBM_{yr-1} + MVN(0, \Sigma)}
}
\examples{

dat = simulate_weight() \%>\%
 mutate(year = as.factor(year)) \%>\%
 arrange(year, age, weight)
# - Plot the data
ggplot(dat, aes(x = age, y = weight, colour = year)) +
 geom_point(size = 2) +
 scale_color_discrete()
}
